pyproject.toml
[project]
name = "paginated-ai-web-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-storage-blob==12.19.0",
    "confluent-kafka>=2.12.2",
    "crawl4ai==0.4.247",
    "pydantic==2.10.6",
    "python-dotenv==1.0.1",
    "requests>=2.32.5",
]
